% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cutData}
\alias{cutData}
\title{cutData}
\usage{
cutData(uncutfiles,
         columns = list("x_coord" = 1, "y_coord" = 2,
                             "z_coord" = 3, "class" = 4),
         omit.class = FALSE)
}
\arguments{
\item{uncutfiles}{a list of raw data e.g. uncut manual measures
columns a list given the column of x-coordinates, y-coordinates, z-coordinates
and class
omit.class logical, if \code{TRUE} a fourth column will be added containing
the assigned class}
}
\value{
a list of data frames with x-coordinates, y-coordinates, z-coordinates
and class (optional).
}
\description{
a function to extract x, y, z and class from raw data.
}
\examples{
\dontrun{
cutData(uncutfiles,
         columns = list("x_coord" = 1, "y_coord" = 2,
                             "z_coord" = 3, "class" = 4),
         omit.class = FALSE)
}
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
}

