\name{DSMcompare-package}
\alias{DSMcompare}
\docType{package}
\title{
Analysis of digital surface models using nonparametric statistics
}
\description{
Compare digital surface models with given manual measures using nonparametric statistics
}
\details{
\tabular{ll}{
Package: \tab DSMcompare\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2014-10-26\cr
License: \tab GPL-2 \cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
Maintainer: Thomas Naake <naake@stud.uni-heidelberg.de>
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package }
\seealso{
  http://www.fva-bw.de/fva/abt.php
  https://github.com/tnaake/DSMcompar
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
