% Generated by roxygen2 (4.0.2): do not edit by hand
\name{errorModel}
\alias{errorModel}
\title{Calculate error values}
\usage{
errorModel(manual, model)
}
\arguments{
\item{manual}{a matrix or data.frame containing x-, y-, z-coordinates and classes
model a list containing model values}
}
\value{
a list of data frames with x-coordinates, y-coordinates, z-coordinates
and class.
}
\description{
a function to calculate the errors of model and manual
measurement values
}
\details{
see Shepard for IDW.
}
\examples{
\dontrun{
errorModel(manual, model)
}
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
}

