% Generated by roxygen2 (4.0.2): do not edit by hand
\name{errorNormalTest}
\alias{errorNormalTest}
\title{Kolmogorov-Smirnow and plot}
\usage{
errorNormalTest(error, hist = TRUE, ksTest = FALSE, qq = FALSE,
  classes = FALSE)
}
\arguments{
\item{error}{a list of errors to compare, each list entry is of class
\code{data.frame} which contains x-coordinates (first column),
y-coordinates (second column), z-coordinates (third column) and
(optionally) classes (fourth column)}
}
\description{
Kolmogorov-Smirnow and plot
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
}

