% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meanManualMeasure}
\alias{meanManualMeasure}
\title{meanManualMeasure}
\usage{
meanManualMeasure(manualMeasure)
}
\arguments{
\item{manualMeasure}{a list of manual measurements, each list entry is of
class \code{data.frame} which contains x-coordinates (first column),
y-coordinates (second column), z-coordinates (third column) and
(optionally) class (fourth column)}
}
\value{
a data.frame containing x-coordinates, y-coordinates, z-coordinates and
class
}
\description{
Calculate mean values of manual measurements.
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
}

