% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readManualMeasure}
\alias{readManualMeasure}
\title{Read manual measurements}
\usage{
readManualMeasure(files)
}
\arguments{
\item{files}{a vector with characters giving the file paths.}
}
\value{
a list which contains the individual models
}
\description{
Read .shp files using maptools::readShapePoints() of a
vector with given file paths pointing to the favoured files. The function
will read the files and create an list object which contains the files.
}
\examples{
\dontrun{
readManualMeasure(files)
}
}
\author{
Thomas Naake <naake@stud.uni-heidelberg.de>
}

