% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stat}
\alias{stat}
\title{Non-parametric statistics for non-normally distributed errors}
\usage{
stat(error, cfi = TRUE, classes = FALSE)
}
\arguments{
\item{error}{A list containing error values of models, each list entry is of class
\code{data.frame} which contains x-coordinates (first column),
y-coordinates (second column), z-coordinates (third column) and
(optionally) classes (fourth column)}

\item{classes}{logical, indicating if classes will be used concerning calculation
of the statistics}

\item{param}{A vector comprising characters which methods to use}
}
\value{
A list. Each list entry comprises a data frame which bears the
statistical parameters
}
\description{
Non-parametric statistics for non-normally distributed errors
}

